+++ temba/settings_common.py
@@ -34,7 +34,7 @@ TEST_EXCLUDE = ("smartmin",)
 # Email
 # -----------------------------------------------------------------------------------
 
-SEND_EMAILS = False
+SEND_EMAILS = TESTING  # enable sending emails in tests
 
 EMAIL_HOST = "smtp.gmail.com"
 EMAIL_HOST_USER = "server@temba.io"
@@ -989,3 +989,7 @@ WHATSAPP_FACEBOOK_BUSINESS_ID = os.environ.get("WHATSAPP_FACEBOOK_BUSINESS_ID", 
 #
 # You need to change these to real addresses to work with these.
 IP_ADDRESSES = ("172.16.10.10", "162.16.10.20")
+
+# Android clients FCM config
+ANDROID_FCM_PROJECT_ID = os.environ.get("ANDROID_FCM_PROJECT_ID", "")
+ANDROID_FCM_SERVICE_ACCOUNT_FILE = os.environ.get("ANDROID_FCM_SERVICE_ACCOUNT_FILE", "")